add shop image & fix bug

+++ bee_core/src/main/java/com/bee/busi/model/order/BusiOrderNumberStat.java
@@ -8,28 +8,28 @@ public class BusiOrderNumberStat implements java.io.Serializable {
     private static final long serialVersionUID = 9090238372407180926L;
 
     // 今天成功订单数
-    private Long today;
+    private Integer today;
     // 今日总订单数
-    private Long todayTotal;
+    private Integer todayTotal;
     // 一周订单数
-    private Long total;
+    private Integer total;
 
-    public Long getToday() {
+    public Integer getToday() {
         return today;
     }
-    public void setToday(Long today) {
+    public void setToday(Integer today) {
         this.today = today;
     }
-    public Long getTodayTotal() {
+    public Integer getTodayTotal() {
         return todayTotal;
     }
-    public void setTodayTotal(Long todayTotal) {
+    public void setTodayTotal(Integer todayTotal) {
         this.todayTotal = todayTotal;
     }
-    public Long getTotal() {
+    public Integer getTotal() {
         return total;
     }
-    public void setTotal(Long total) {
+    public void setTotal(Integer total) {
         this.total = total;
     }
 }
